Refactor: entity 패키지 추가
