+++ notification/src/main/java/org/bookwoori/notification/domain/entity/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package org.bookwoori.notification.domain;
+package org.bookwoori.notification.domain.entity;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
