+++ notification/src/main/java/org/bookwoori/notification/domain/entity/Notification.java
@@ -1,4 +1,4 @@
-package org.bookwoori.notification.domain;
+package org.bookwoori.notification.domain.entity;
 
 import lombok.*;
 import org.springframework.data.annotation.Id;
