+++ notification/src/main/java/org/bookwoori/notification/domain/repository/DeviceRepository.java
@@ -1,7 +1,7 @@
 package org.bookwoori.notification.domain.repository;
 
 
-import org.bookwoori.notification.domain.Device;
+import org.bookwoori.notification.domain.entity.Device;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
