+++ notification/src/main/java/org/bookwoori/notification/domain/service/NotificationService.java
@@ -7,8 +7,8 @@ import com.google.firebase.messaging.FirebaseMessagingException;
 import com.google.firebase.messaging.MulticastMessage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.bookwoori.notification.domain.Device;
-import org.bookwoori.notification.domain.Notification;
+import org.bookwoori.notification.domain.entity.Device;
+import org.bookwoori.notification.domain.entity.Notification;
 import org.bookwoori.notification.domain.dto.request.ChannelMessageRequestDto;
 import org.bookwoori.notification.domain.dto.request.ChatMessageRequestDto;
 import org.bookwoori.notification.domain.dto.request.DirectMessageRequestDto;
